+++ backend/src/app.ts
@@ -5,8 +5,8 @@ import "./db";
 import dashboardRouter from "./routes/index";
 import usersRouter from "./routes/Users.routes";
 import projectsRouter from "./routes/Projects.routes";
-import projectPostRouter from "./routes/ProjectPosts.routes";
-import projectPostReplyRouter from "./routes/ProjectPostReply";
+import projectPostRouter from "./routes/ProjectsPosts";
+import projectPostReplyRouter from "./routes/ProjectsPostsReply";
 import githubRouter from "./routes/Auth.routes";
 import friendshipsRouter from "./routes/Friendship";
 import userPostsRouter from "./routes/UserPosts.routes";
